+++ src/components/map/filters/settings/MapSettingsPresets.vue
@@ -15,7 +15,7 @@
                     Save Current Settings
                 </common-block-title>
 
-                <div class="presets__row">
+                <div class="presets__row presets__row--no-wrap">
                     <common-input-text
                         v-model="newPresetName"
                         placeholder="Preset Name"
@@ -46,7 +46,10 @@
                     Saved Presets
                 </common-block-title>
 
-                <div class="__grid-info-sections">
+                <div
+                    class="__grid-info-sections"
+                    :class="{ '__grid-info-sections--large-title': isMobile }"
+                >
                     <div class="__grid-info-sections_title">
                         Preset
                     </div>
@@ -121,7 +124,10 @@
                 </div>
 
                 <template v-if="activePreset">
-                    <div class="__grid-info-sections">
+                    <div
+                        class="__grid-info-sections"
+                        :class="{ '__grid-info-sections--large-title': isMobile }"
+                    >
                         <div class="__grid-info-sections_title">
                             Rename Preset
                         </div>
@@ -250,6 +256,7 @@ const presets = computed(() => store.mapPresets);
 
 const newPresetName = ref('');
 const activePreset = shallowRef<UserMapPreset | null>(null);
+const isMobile = useIsMobile();
 
 const currentPreset = computed(() => {
     return presets.value.find(x => JSON.stringify(x.json) === JSON.stringify(store.mapSettings))?.id ?? null;
@@ -327,6 +334,13 @@ const presetsSelectList = computed<SelectItem[]>(() => presets.value.map(x => ({
         gap: 16px;
         align-items: center;
 
+        &:not(.presets__row--no-wrap) {
+            @include mobileOnly {
+                flex-wrap: wrap;
+                row-gap: 8px;
+            }
+        }
+
         .button {
             min-width: 32px;
         }
@@ -335,6 +349,10 @@ const presetsSelectList = computed<SelectItem[]>(() => presets.value.map(x => ({
             width: 1px;
             height: 24px;
             background: varToRgba('lightgray150', 0.2);
+
+            @include mobileOnly {
+                display: none;
+            }
         }
     }