+++ src/components/map/popups/MapPopupAtc.vue
@@ -233,7 +233,9 @@ const { data: stats } = useLazyAsyncData(`stats-atc-${ atc.value?.cid ?? Math.ra
 
 <style scoped lang="scss">
 .atc {
-    height: 100%;
+    @include fromTablet {
+        height: 100%;
+    }
 
     &__sections {
         display: flex;