+++ src/components/views/ViewMapFooter.vue
@@ -44,50 +44,7 @@
                         Featured Airports
                     </template>
 
-                    <div class="__info-sections map-footer__featured-content">
-                        <common-tabs
-                            v-model="featuredTab"
-                            :tabs="{
-                                popular: {
-                                    title: 'Popular Airports',
-                                },
-                                quiet: {
-                                    title: 'Quiet Airports',
-                                },
-                            }"
-                        />
-
-                        <small v-if="featuredTab === 'quiet'">
-                            Quiet, staffed airports with at least one controller available, excluding center.
-                        </small>
-
-                        <div class="__section-group">
-                            <common-toggle
-                                :model-value="store.localSettings.traffic?.showTotalDeparturesInFeaturedAirports ?? false"
-                                @update:modelValue="setUserLocalSettings({ traffic: { showTotalDeparturesInFeaturedAirports: $event } })"
-                            >
-                                Show total departures
-                                <template #description>
-                                    Including airborne
-                                </template>
-                            </common-toggle>
-                            <common-toggle v-model="store.featuredVisibleOnly">
-                                Visible only
-                                <template #description>
-                                    Filter by current map area
-                                </template>
-                            </common-toggle>
-                        </div>
-
-                        <div class="__info-sections">
-                            <common-airport-card
-                                v-for="(airport, index) in (featuredTab === 'popular' ? popularAirports : quietAirports)"
-                                :key="airport.airport.icao + index"
-                                :airport="airport"
-                                :position="index + 1"
-                            />
-                        </div>
-                    </div>
+                    <map-featured-airports/>
                 </common-control-block>
             </div>
             <div class="map-footer_left_section __desktop">
@@ -140,11 +97,11 @@
             </div>
         </div>
         <div
-            v-if="getLastUpdated"
+            v-if="getCounts.lastUpdated"
             class="map-footer_right __from-tablet"
             :class="{ 'map-footer_right--outdated': outdated }"
         >
-            Map last updated: {{ getLastUpdated }}
+            Map last updated: {{ getCounts.lastUpdated }}
         </div>
     </footer>
     <common-popup v-model="airacPopup">
@@ -209,53 +166,16 @@
 import { useStore } from '~/store';
 import CommonButton from '~/components/common/basic/CommonButton.vue';
 import CommonControlBlock from '~/components/common/blocks/CommonControlBlock.vue';
-import CommonAirportCard from '~/components/common/vatsim/CommonAirportCard.vue';
-import CommonTabs from '~/components/common/basic/CommonTabs.vue';
-import CommonToggle from '~/components/common/basic/CommonToggle.vue';
 import { useOnlineCounters } from '~/composables/navigation';
+import MapFeaturedAirports from '~/components/map/MapFeaturedAirports.vue';
 
 const store = useStore();
 const dataStore = useDataStore();
 const airacPopup = ref(false);
-const featuredTab = ref('popular');
-
-const datetime = new Intl.DateTimeFormat([], {
-    timeZone: 'UTC',
-    localeMatcher: 'best fit',
-    hour: '2-digit',
-    minute: '2-digit',
-    second: '2-digit',
-});
 
 const getCounts = useOnlineCounters();
 
-const getLastUpdated = computed(() => {
-    const updateTimestamp = dataStore.vatsim.updateTime.value;
-    if (!updateTimestamp) return null;
-
-    const date = new Date(updateTimestamp);
-
-    return `${ datetime.format(date) } Z`;
-});
-
 const outdated = computed(() => dataStore.time.value - dataStore.vatsim.updateTime.value > 1000 * 20);
-
-const popularAirports = computed(() => {
-    return dataStore.vatsim.parsedAirports.value.filter(x => !x.airport.isPseudo && x.aircraftCids.length).slice().sort((a, b) => b.aircraftCids.length - a.aircraftCids.length).slice(0, store.featuredVisibleOnly ? 10 : 25);
-});
-
-const quietAirports = computed(() => {
-    return dataStore.vatsim.parsedAirports.value
-        .filter(x => !x.airport.isPseudo && (x.aircraftCids.length || x.localAtc.some(x => x.isATIS)) && (x.arrAtc.length || x.localAtc.some(x => !x.isATIS)))
-        .slice()
-        .sort((a, b) => {
-            const aSum = (a.aircraftList.arrivals?.length ?? 0) + (a.aircraftList.groundDep?.length ?? 0);
-            const bSum = (b.aircraftList.arrivals?.length ?? 0) + (b.aircraftList.groundDep?.length ?? 0);
-
-            return aSum - bSum;
-        })
-        .slice(0, 25);
-});
 </script>
 
 <style scoped lang="scss">
@@ -267,10 +187,6 @@ const quietAirports = computed(() => {
     padding: 0 24px;
 
     font-size: 13px;
-
-    & &__featured-content {
-        gap: 8px;
-    }
 
     &_left {
         display: flex;